+++ src/main/java/com/samaxes/maven/minify/plugin/ProcessFilesTask.java
@@ -28,7 +28,6 @@ import java.io.InputStreamReader;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
 import java.io.SequenceInputStream;
-import java.net.URI;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -51,23 +50,23 @@ public abstract class ProcessFilesTask implements Callable<Object> {
 
     protected Log log;
 
-    protected Integer bufferSize;
-
     protected boolean debug;
+
+    protected Integer bufferSize;
 
     protected boolean skipMerge;
 
     protected boolean skipMinify;
 
-    private String mergedFilename;
-
-    private String suffix;
-
-    protected boolean nosuffix;
-
     protected String charset;
 
     protected int linebreak;
+
+    private String suffix;
+
+    private boolean nosuffix;
+
+    private String mergedFilename;
 
     private File sourceDir;
 
@@ -98,9 +97,9 @@ public abstract class ProcessFilesTask implements Callable<Object> {
      * @param outputDir directory to write the final file
      * @param outputFilename the output file name
      * @param suffix final filename suffix
-     * @param nosuffix whether to use a suffix for the minified file name or not
+     * @param nosuffix whether to use a suffix for the minified filename or not
      * @param charset if a character set is specified, a byte-to-char variant allows the encoding to be selected.
-     * Otherwise, only byte-to-byte operations are used
+     *        Otherwise, only byte-to-byte operations are used
      * @param linebreak split long lines after a specific column
      */
     public ProcessFilesTask(Log log, Integer bufferSize, boolean debug, boolean skipMerge, boolean skipMinify,
@@ -108,8 +107,8 @@ public abstract class ProcessFilesTask implements Callable<Object> {
             List<String> sourceIncludes, List<String> sourceExcludes, String outputDir, String outputFilename,
             String suffix, boolean nosuffix, String charset, int linebreak) {
         this.log = log;
-        this.bufferSize = bufferSize;
         this.debug = debug;
+        this.bufferSize = bufferSize;
         this.skipMerge = skipMerge;
         this.skipMinify = skipMinify;
         this.mergedFilename = outputFilename;
@@ -150,11 +149,9 @@ public abstract class ProcessFilesTask implements Callable<Object> {
                             originalPath.lastIndexOf(File.separator));
                     File targetPath = new File(targetDir.getAbsolutePath() + subPath);
                     targetPath.mkdirs();
-                    log.info("nosuffix?" + nosuffix);
 
                     File minifiedFile = new File(targetPath, (nosuffix) ? mergedFile.getName() : mergedFile.getName()
                             .replace(extension, suffix + extension));
-                    log.info("Calling minify on: " + minifiedFile.getName());
                     minify(mergedFile, minifiedFile);
                 }
             } else if (skipMinify) {
@@ -163,9 +160,9 @@ public abstract class ProcessFilesTask implements Callable<Object> {
                 log.info("Skipping minify step.");
             } else {
                 File mergedFile = new File(targetDir, (nosuffix) ? mergedFilename + TEMP_SUFFIX : mergedFilename);
-                File minifiedFile = new File(targetDir, (nosuffix) ? mergedFilename
-                        : mergedFile.getName().replace(extension, suffix + extension));
                 merge(mergedFile);
+                File minifiedFile = new File(targetDir, (nosuffix) ? mergedFilename : mergedFile.getName().replace(
+                        extension, suffix + extension));
                 minify(mergedFile, minifiedFile);
                 if (nosuffix) {
                     if (!mergedFile.delete()) {
@@ -221,7 +218,7 @@ public abstract class ProcessFilesTask implements Callable<Object> {
      * @param mergedFile input file resulting from the merged step
      * @param minifiedFile output file resulting from the minify step
      */
-    abstract void minify(File file, File minifiedFile);
+    abstract void minify(File mergedFile, File minifiedFile);
 
     /**
      * Logs an addition of a new source file.
@@ -233,27 +230,6 @@ public abstract class ProcessFilesTask implements Callable<Object> {
         File sourceFile = new File(sourceDir, sourceFilename);
 
         addNewSourceFile(finalFilename, sourceFile);
-    }
-
-    /**
-     * Cleanup files. Remove merged file is nosuffix parameter was set to true.
-     *
-     * @param mergedFile input file resulting from the merged step
-     * @param minifiedFile output file resulting from the minify step
-     */
-    void cleanupFiles(File mergedFile, File minifiedFile) {
-        if (nosuffix) {
-            String mergedFilename = mergedFile.getName();
-            URI mergedFileURI = mergedFile.toURI();
-            if (debug) {
-                log.info("Deleting the file [" + mergedFilename + "].");
-            }
-            mergedFile.delete();
-            if (debug) {
-                log.info("Renaming the file [" + minifiedFile.getName() + "] to [" + mergedFilename + "].");
-            }
-            minifiedFile.renameTo(new File(mergedFileURI));
-        }
     }
 
     /**