+++ src/main/java/com/samaxes/maven/minify/plugin/ProcessJSFilesTask.java
@@ -75,7 +75,7 @@ public class ProcessJSFilesTask extends ProcessFilesTask {
      * @param outputDir directory to write the final file
      * @param outputFilename the output file name
      * @param suffix final filename suffix
-     * @param nosuffix whether to use a suffix for the minified file name or not
+     * @param nosuffix whether to use a suffix for the minified filename or not
      * @param charset if a character set is specified, a byte-to-char variant allows the encoding to be selected.
      *        Otherwise, only byte-to-byte operations are used
      * @param linebreak split long lines after a specific column
@@ -87,8 +87,8 @@ public class ProcessJSFilesTask extends ProcessFilesTask {
     public ProcessJSFilesTask(Log log, Integer bufferSize, boolean debug, boolean skipMerge, boolean skipMinify,
             String webappSourceDir, String webappTargetDir, String inputDir, List<String> sourceFiles,
             List<String> sourceIncludes, List<String> sourceExcludes, String outputDir, String outputFilename,
-            String suffix, boolean nosuffix, String charset, int linebreak, String jsEngine, boolean munge, boolean verbose,
-            boolean preserveAllSemiColons, boolean disableOptimizations) {
+            String suffix, boolean nosuffix, String charset, int linebreak, String jsEngine, boolean munge,
+            boolean verbose, boolean preserveAllSemiColons, boolean disableOptimizations) {
         super(log, bufferSize, debug, skipMerge, skipMinify, webappSourceDir, webappTargetDir, inputDir, sourceFiles,
                 sourceIncludes, sourceExcludes, outputDir, outputFilename, suffix, nosuffix, charset, linebreak);
 
@@ -106,9 +106,12 @@ public class ProcessJSFilesTask extends ProcessFilesTask {
      * @param minifiedFile output file resulting from the minify step
      */
     @Override
-    protected void minify(File mergedFile, File minifiedFile) {
+    void minify(File mergedFile, File minifiedFile) {
         if (minifiedFile != null) {
             try {
+                log.info("Creating minified file [" + ((debug) ? minifiedFile.getPath() : minifiedFile.getName())
+                        + "].");
+
                 InputStream in = new FileInputStream(mergedFile);
                 OutputStream out = new FileOutputStream(minifiedFile);
                 InputStreamReader reader;
@@ -120,14 +123,7 @@ public class ProcessJSFilesTask extends ProcessFilesTask {
                     reader = new InputStreamReader(in, charset);
                     writer = new OutputStreamWriter(out, charset);
                 }
-                
-                if (debug) {
-                    log.info("Creating minified file [" + minifiedFile.getPath() + "].");
-                } else {
-                    File temp = (nosuffix) ? mergedFile : minifiedFile;
-                    log.info("Creating minified file [" + temp.getName() + "].");
-                }
-                
+
                 if ("closure".equals(jsEngine)) {
                     log.debug("Using JavaScript compressor engine [Google Closure Compiler].");
 
@@ -157,9 +153,6 @@ public class ProcessJSFilesTask extends ProcessFilesTask {
             } catch (IOException e) {
                 log.error(e.getMessage(), e);
             }
-
-            cleanupFiles(mergedFile, minifiedFile);
-
         }
     }
 }