+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Minify Maven Plugin
 
+## 1.6.2
+
+* Cherry picked from commit 31fe5c91bf2d24c29251595206c3c4ebada1c712: [MINIFY-16] Added nosuffix option to avoid the suffix (.min) on the minified output filename.
+
 ## 1.6.1
 
-* Preserve sub-directory structure when only minifying (#29).
+* Cherry picked from commit 924a23a373e6b9aa841af6b9e4300c670eb602aa: Preserve sub-directory structure when only minifying (#29).
 
 ## 1.6